Tests/LibWeb/Text/input/Editing/clipboard-shortcut-iframe.html

<!DOCTYPE html>
<script src="../include.js"></script>
<iframe id="frame" srcdoc="<body contenteditable>blue text</body>"></iframe>
<script>
    asyncTest(async done => {
        await new Promise(resolve => frame.addEventListener("load", resolve, { once: true }));

        const platformControl = internals.MOD_CTRL | internals.MOD_SUPER;
        const body = frame.contentDocument.body;
        const selection = frame.contentWindow.getSelection();
        selection.selectAllChildren(body);

        body.addEventListener("copy", event => {
            event.clipboardData.setData("text/plain", "blue text");
            event.clipboardData.setData("text/html", '<span style="color: blue">blue text</span>');
            event.preventDefault();
        }, { once: true });
        internals.sendKey(body, "C", platformControl);
        internals.dispatchUserActivatedEvent(body, new Event("mousedown"));
        const [clipboardItem] = await frame.contentWindow.navigator.clipboard.read();
        println(`copy formats: ${JSON.stringify(clipboardItem.types)}`);

        const pasteEvent = new Promise(resolve => body.addEventListener("paste", event => {
            event.preventDefault();
            println(`paste plain: ${JSON.stringify(event.clipboardData.getData("text/plain"))}`);
            println(`paste html: ${JSON.stringify(event.clipboardData.getData("text/html"))}`);
            resolve();
        }, { once: true }));
        selection.collapse(body.firstChild, body.firstChild.length);
        internals.sendKey(body, "V", platformControl);
        await pasteEvent;
        done();
    });
</script>
