Tests/LibWeb/Text/expected/Editing/clipboard-shortcut-iframe.txt

copy formats: ["text/plain","text/html"]
paste plain: "blue text"
paste html: "<span style=\"color: blue\">blue text</span>"
